dedicated --user test suite; use system_site_packages option for reset_env
